% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-utils.R
\name{convert_codecat}
\alias{convert_codecat}
\title{Converts and renames QA data \verb{[code]} and \verb{[code]_codetype} columns to \code{code}
and \code{codecat} in a data frame}
\usage{
convert_codecat(
  data,
  codecat = c("Rx", "rx", "RX", "PX", "Px", "px", "DX", "Dx", "dx")
)
}
\arguments{
\item{data}{a data frame}

\item{codecat}{a character vector of codecat values (default:
c("Rx","rx","RX","PX","Px","px","DX","Dx","dx"))}
}
\value{
a data frame with renamed \code{codecat} column
}
\description{
Converts and renames QA data \verb{[code]} and \verb{[code]_codetype} columns to \code{code}
and \code{codecat} in a data frame
}
