% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_orphan_enc}
\alias{create_orphan_enc}
\title{Get counts of orphaned and unorphaned EncounterIDs.}
\usage{
create_orphan_enc(data, linked.tables = NULL, cols = NULL)
}
\arguments{
\item{data}{Data frame containing the variables \code{enc}, and \code{count}.}

\item{linked.tables}{Character vector of tables to be used
in the calculation of \code{matched}.}

\item{cols}{Columns to group by in the calculation of utilization summary.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Creates a table displaying the counts of EncounterIDs found the ENC table,
and returns TRUE if those EncounterIDs have a match in any or both of the DIA
and PRO tables, and FALSE if they do not.
}
