% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{load_data}
\alias{load_data}
\title{Load data from a SAS file}
\usage{
load_data(source, file = NULL)
}
\arguments{
\item{source}{The source folder of the file}

\item{file}{The name of the file}
}
\value{
A data frame read from the SAS file
}
\description{
Load data from a SAS file
}
