% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{safely_errors}
\alias{safely_errors}
\title{Detect and print errors returned by an invocation of \code{\link[purrr:safely]{purrr::safely()}}}
\usage{
safely_errors(.x)
}
\arguments{
\item{.x}{A list containing the results of an invocation of \code{\link[purrr:safely]{purrr::safely()}}}
}
\description{
Detect and print errors returned by an invocation of \code{\link[purrr:safely]{purrr::safely()}}
}
