% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-utils.R
\name{subset_list}
\alias{subset_list}
\title{Subset a list based on search term}
\usage{
subset_list(.x, search_term)
}
\arguments{
\item{.x}{A list}

\item{search_term}{The search term to use for subsetting the list}
}
\value{
A subset of the input list, containing only elements whose names
contain the search term
}
\description{
Subset a list based on search term
}
