% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-utils.R
\name{subset_tabid}
\alias{subset_tabid}
\title{Subset a list based on tabid term}
\usage{
subset_tabid(.x, tabid)
}
\arguments{
\item{.x}{A list}

\item{tabid}{The three letter tabid used to subset the list}
}
\value{
A subset of the input list, containing only elements whose names
start with the search term
}
\description{
Subset a list based on tabid term
}
