% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_match.R
\name{tidy_match}
\alias{tidy_match}
\title{Convert table variables in match datasets to logicals}
\usage{
tidy_match(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a tidied data frame
}
\description{
Tidy the columns of data frame in such a way that the values in dp and count
columns are recoded to be logical, i.e. 0 becomes FALSE and 1 becomes TRUE.
}
