% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_utilization}
\alias{create_utilization}
\title{Get counts of utilizing and non-utilizing PatIDs.}
\usage{
create_utilization(data, utilization.tables = NULL, cols = NULL)
}
\arguments{
\item{data}{Data frame containing the variables \code{dem}, \code{enr}, and \code{count}.}

\item{utilization.tables}{Character vector of utilization tables to be used
in the calculation of \code{utilizing}.}

\item{cols}{Columns to group by in the calculation of utilization summary.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Creates a table displaying the counts of PatIDs found in both the DEM and ENR
tables and returns TRUE if those PatIDs are found in any of the DIA, DIS,
DTH, ENC, PRO or only one of the DEM & ENR tables.
}
\examples{
create_utilization(data = mydata, utilization.tables = c("DIA", "DIS", "DTH", "ENC", "PRO"))
create_utilization(data = mydata, utilization.tables = c("DIA", "DIS", "DTH", "ENC", "PRO"), cols = c("dp","jurisdiction"))

}
