% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelTidyData-package.R
\docType{package}
\name{sentinelTidyData-package}
\alias{sentinelTidyData}
\alias{sentinelTidyData-package}
\title{sentinelTidyData: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel Scarnecchia \email{scarnecchia@gmail.com}

}
\keyword{internal}
